refactor: consolidate customs services
