+++ bot_alista/services/currency.py
@@ -17,22 +17,33 @@ def _get_converter() -> CurrencyConverter:
         _converter = CurrencyConverter()
     return _converter
 
-
-def to_eur(amount: float, currency: str) -> float:
+def to_eur(amount: float, currency: str, eur_rate: float | None = None) -> float:
     """Convert ``amount`` from ``currency`` to EUR.
+
+    ``eur_rate`` may be provided to convert values expressed in RUB without
+    contacting the external rate provider.  When omitted the function attempts
+    to obtain the rate from :mod:`currency_converter_free` and finally falls
+    back to the static table.
 
     :param amount: value in the original currency
     :param currency: ISO currency code, e.g. ``"USD"``
+    :param eur_rate: optional EUR/RUB rate for deterministic RUB conversion
     :return: amount converted to euros
     :raises ValueError: if currency is unsupported
     """
+
     code = currency.upper()
     if code == "EUR":
         return float(amount)
+    if code == "RUB" and eur_rate is not None:
+        return float(amount) / eur_rate
+
     try:
         converter = _get_converter()
         return float(converter.convert(amount, code, "EUR"))
     except Exception:
+        if code == "RUB" and eur_rate is not None:
+            return float(amount) / eur_rate
         rate = _FALLBACK_RATES.get(code)
         if rate is None:
             raise ValueError(f"Unsupported currency: {currency}")
